Helper function for constructing items for CommandBar

CommandBar expects items definition as a nested structure, which gets lengthy and verbose. This function helps makes this definition shorter. Returns a list with all arguments passed through, except for text, icon (which will inserted as proper iconProps) and subitems (which will be inserted as proper subMenuProps)

CommandBarItem(text, icon = NULL, subitems = NULL, ...)

Arguments

text

text to be displayed on the menu

icon

optional name of an icon

subitems

optional list of CommandBar items

See also