Fluent UI themes

Themes is not a component, but a way to override the default theme in the app. See https://developer.microsoft.com/en-us/fluentui#/controls/web/themes for details.

Examples

library(shiny.fluent)

if (interactive()) {
  shinyApp(
    ui = tagList(
      tags$script(HTML("
Fluent = window['shiny.fluent']
Fluent.loadTheme({
  defaultFontStyle: { fontFamily: 'Times New Roman', fontWeight: 'regular' }
})
")),
      withReact(
        Label('The default theme font has been changed to Times New Roman in this example')
      )
    ),
    server = function(input, output) {
    }
  )
}