A dataset of sample people based on Fluent UI examples

A dataset of sample people based on Fluent UI examples

Usage

fluentPeople

Format

An object of class tbl_df (inherits from tbl, data.frame) with 7 rows and 11 columns.

Source

https://developer.microsoft.com/en-us/fluentui#/controls/web/peoplepicker